backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.2.0
numpy==1.26.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
greenlet==3.3.1
